refactor(auth): replace OAuthLauncher with a cookie-only store
