+++ feature/auth/build.gradle.kts
@@ -11,9 +11,6 @@ kotlin {
         commonMain.dependencies {
             implementation(project(":core:network"))
         }
-        androidMain.dependencies {
-            implementation(libs.browser)
-        }
         named("androidUnitTest").dependencies {
             implementation(libs.kotlinx.serialization.json)
             implementation(libs.koin.test)
