+++ feature/auth/src/commonMain/kotlin/com/garfiec/librechat/feature/auth/oauth/OAuthCookieStore.kt
@@ -1,0 +1,7 @@
+package com.garfiec.librechat.feature.auth.oauth
+
+/** Platform cookie-jar access for the SSO WebView flow: wipes the server's stale `refreshToken`
+ *  cookie before an SSO round-trip so only a cookie minted by that round-trip can be captured. */
+interface OAuthCookieStore {
+    fun clearRefreshTokenCookie(serverUrl: String)
+}
