+++ feature/auth/src/commonMain/kotlin/com/garfiec/librechat/feature/auth/viewmodel/LoginViewModel.kt
@@ -5,12 +5,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.garfiec.librechat.core.common.result.ApiException
 import com.garfiec.librechat.core.common.result.Result
-import com.garfiec.librechat.core.data.datastore.ServerDataStore
 import com.garfiec.librechat.core.data.repository.AccountSwitcher
 import com.garfiec.librechat.core.data.repository.AuthRepository
 import com.garfiec.librechat.core.data.repository.ConfigRepository
 import com.garfiec.librechat.core.model.LoginOutcome
-import com.garfiec.librechat.feature.auth.oauth.OAuthLauncher
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -36,8 +34,6 @@ data class LoginUiState(
 class LoginViewModel(
     private val authRepository: AuthRepository,
     private val configRepository: ConfigRepository,
-    private val oAuthLauncher: OAuthLauncher,
-    private val serverDataStore: ServerDataStore,
     private val accountSwitcher: AccountSwitcher,
 ) : ViewModel() {
 
@@ -63,10 +59,6 @@ class LoginViewModel(
             }
         }
     }
-
-    /** The server this screen is signing into: the pending add target when set, else the live one. */
-    private fun signInServerUrl(): String =
-        accountSwitcher.pendingAdd?.serverUrl ?: serverDataStore.getBaseUrl()
 
     fun onEmailChanged(email: String) {
         _uiState.value = _uiState.value.copy(email = email, error = null)
@@ -130,59 +122,5 @@ class LoginViewModel(
 
     fun consumeTwoFactorNavigation() {
         _uiState.value = _uiState.value.copy(twoFactorTempToken = null)
-    }
-
-    /** Set once this screen launches its own OAuth round-trip; gates add-mode cookie consumption. */
-    private var oAuthLaunched = false
-
-    fun launchOAuth(provider: String) {
-        oAuthLaunched = true
-        val serverUrl = signInServerUrl()
-        // Drop any stale refreshToken cookie for this host BEFORE launching. In add mode the cookie
-        // jar is process-global and nothing clears it on add-flow entry, so a launch that the user then
-        // cancels would otherwise leave a pre-existing cookie for checkOAuthResult() to consume as the
-        // wrong user (the oAuthLaunched guard only blocks the never-launched case). Clearing here means
-        // only a cookie minted by THIS round-trip can be present on return.
-        oAuthLauncher.clearOAuthCookie(serverUrl)
-        oAuthLauncher.launchOAuth(provider, serverUrl)
-    }
-
-    fun checkOAuthResult() {
-        // In add mode, only consume a cookie minted by THIS screen's own launchOAuth round-trip:
-        // the cookie jar is process-global and nothing clears it on add-flow entry, so a stale
-        // refreshToken cookie for this host would otherwise be auto-consumed on first ON_RESUME
-        // and silently complete the add as the wrong user. The normal login screen keeps the
-        // unconditional consume — it must survive process death during the Custom Tab round-trip,
-        // which an add flow never does (its pending session is memory-only, so a killed add flow
-        // is stripped by the NavHost, not resumed).
-        if (accountSwitcher.pendingAdd != null && !oAuthLaunched) return
-
-        val serverUrl = signInServerUrl()
-        if (serverUrl.isBlank()) return
-
-        val refreshToken = oAuthLauncher.extractTokenFromCookies(serverUrl) ?: return
-
-        // Clear the cookie immediately to avoid re-reading on next onResume
-        oAuthLauncher.clearOAuthCookie(serverUrl)
-
-        viewModelScope.launch {
-            _uiState.value = _uiState.value.copy(isLoading = true, error = null)
-
-            when (val result = authRepository.loginWithOAuthToken(refreshToken)) {
-                is Result.Success -> {
-                    _uiState.value = _uiState.value.copy(
-                        isLoading = false,
-                        isLoggedIn = true,
-                    )
-                }
-                is Result.Error -> {
-                    _uiState.value = _uiState.value.copy(
-                        isLoading = false,
-                        error = result.message ?: "OAuth login failed",
-                    )
-                }
-                is Result.Loading -> { /* no-op */ }
-            }
-        }
     }
 }
