+++ feature/auth/src/iosMain/kotlin/com/garfiec/librechat/feature/auth/di/AuthPlatformModule.kt
@@ -1,10 +1,10 @@
 package com.garfiec.librechat.feature.auth.di
 
-import com.garfiec.librechat.feature.auth.oauth.IosOAuthLauncher
-import com.garfiec.librechat.feature.auth.oauth.OAuthLauncher
+import com.garfiec.librechat.feature.auth.oauth.IosOAuthCookieStore
+import com.garfiec.librechat.feature.auth.oauth.OAuthCookieStore
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
 actual val authPlatformModule: Module = module {
-    single<OAuthLauncher> { IosOAuthLauncher() }
+    single<OAuthCookieStore> { IosOAuthCookieStore() }
 }
